requirements.txt
fastapi
uvicorn
python-multipart
jinja2
pypdf
google-genai
Python_dotenv
pymongo[srv]==3.12
datetime
python-docx
python-pptx
pandas
xlrd
openpyxl
ebooklib
pdf2image
easyocr
bs4